Updated the delete function in the connections for the delete route so the user has to be verified

+++ Sources/App/Controllers/ConnectionController.swift
@@ -87,9 +87,17 @@ final class ConnectionController {
         throw Abort.notFound
     }
     
-    try connection.delete()
+    //Check if the user requesting the update is equal to the connection user_id
+    if request.headers["user_id"]?.int == connection.invitee_id.int || connection.inviter_id.int {
     
-    return try JSON(node: ["message": "Connection removed."])
-    
+      //Delete the connection
+      try connection.delete()
+
+      //Return a JSON confirmation message
+      return try JSON(node: ["message": "Connection removed."])
+
+    } else {
+      
+    }
   }
 }
